+++ lib/src/react_client/component_registration_utils.dart
@@ -1,0 +1,92 @@
+import 'package:react/react.dart';
+import 'package:react/react_client/bridge.dart';
+import 'package:react/react_client/js_backed_map.dart';
+import 'package:react/react_client/react_interop.dart';
+import 'package:react/react_client/react_proxies.dart';
+
+import 'package:react/src/react_client/react_statics.dart';
+import 'package:react/src/react_client/utils.dart';
+
+/// Creates and returns a new [ReactDartComponentFactoryProxy] from the provided [componentFactory]
+/// which produces a new JS [`ReactClass` component class](https://facebook.github.io/react/docs/top-level-api.html#react.createclass).
+@Deprecated('6.0.0')
+ReactDartComponentFactoryProxy registerComponent(
+    ComponentFactory componentFactory, [
+      Iterable<String> skipMethods = const ['getDerivedStateFromError', 'componentDidCatch'],
+    ]) {
+  var componentInstance = componentFactory();
+
+  if (componentInstance is Component2) {
+    return registerComponent2(componentFactory, skipMethods: skipMethods);
+  }
+
+  var componentStatics = new ComponentStatics(componentFactory);
+
+  var jsConfig = new JsComponentConfig(
+    childContextKeys: componentInstance.childContextKeys,
+    contextKeys: componentInstance.contextKeys,
+  );
+
+  /// Create the JS [`ReactClass` component class](https://facebook.github.io/react/docs/top-level-api.html#react.createclass)
+  /// with custom JS lifecycle methods.
+  var reactComponentClass = createReactDartComponentClass(dartInteropStatics, componentStatics, jsConfig)
+  // ignore: invalid_use_of_protected_member
+    ..dartComponentVersion = ReactDartComponentVersion.component
+    ..displayName = componentFactory().displayName;
+
+  // Cache default props and store them on the ReactClass so they can be used
+  // by ReactDartComponentFactoryProxy and externally.
+  final Map defaultProps = new Map.unmodifiable(componentInstance.getDefaultProps());
+  reactComponentClass.dartDefaultProps = defaultProps;
+
+  return new ReactDartComponentFactoryProxy(reactComponentClass);
+}
+
+
+/// Creates and returns a new [ReactDartComponentFactoryProxy] from the provided [componentFactory]
+/// which produces a new JS [`ReactClass` component class](https://facebook.github.io/react/docs/top-level-api.html#react.createclass).
+ReactDartComponentFactoryProxy2 registerComponent2(
+    ComponentFactory<Component2> componentFactory, {
+      Iterable<String> skipMethods = const ['getDerivedStateFromError', 'componentDidCatch'],
+      Component2BridgeFactory bridgeFactory,
+    }) {
+  bridgeFactory ??= Component2BridgeImpl.bridgeFactory;
+
+  final componentInstance = componentFactory();
+  final componentStatics = new ComponentStatics2(
+    componentFactory: componentFactory,
+    instanceForStaticMethods: componentInstance,
+    bridgeFactory: bridgeFactory,
+  );
+  final filteredSkipMethods = filterSkipMethods(skipMethods);
+
+  // Cache default props and store them on the ReactClass so they can be used
+  // by ReactDartComponentFactoryProxy and externally.
+  final JsBackedMap defaultProps = new JsBackedMap.from(componentInstance.defaultProps);
+
+  final JsMap jsPropTypes =
+  bridgeFactory(componentInstance).jsifyPropTypes(componentInstance, componentInstance.propTypes);
+
+  var jsConfig2 = new JsComponentConfig2(
+    defaultProps: defaultProps.jsObject,
+    contextType: componentInstance.contextType?.jsThis,
+    skipMethods: filteredSkipMethods,
+    propTypes: jsPropTypes,
+  );
+
+  /// Create the JS [`ReactClass` component class](https://facebook.github.io/react/docs/top-level-api.html#react.createclass)
+  /// with custom JS lifecycle methods.
+  var reactComponentClass =
+  createReactDartComponentClass2(ReactDartInteropStatics2.staticsForJs, componentStatics, jsConfig2)
+    ..displayName = componentInstance.displayName;
+  // ignore: invalid_use_of_protected_member
+  reactComponentClass.dartComponentVersion = ReactDartComponentVersion.component2;
+
+  return new ReactDartComponentFactoryProxy2(reactComponentClass);
+}
+
+/// Creates and returns a new `ReactDartFunctionComponentFactoryProxy` from the provided [dartFunctionComponent]
+/// which produces a new `JsFunctionComponent`.
+ReactDartFunctionComponentFactoryProxy registerFunctionComponent(DartFunctionComponent dartFunctionComponent,
+    {String displayName}) =>
+    ReactDartFunctionComponentFactoryProxy(dartFunctionComponent, displayName: displayName);
